Fix details in share permossion
- factor out share advanced permission
- fix note textview & keyboard
  - use (scrollable) textview inside scrollview to enable users to type multiline, while not hiding the header in landscape
- fix footer safe area & padding & dynalic button title
- use NCShareCommon
- remove gesture delegate
